+++ src/main/java/bowling/ui/Request.java
@@ -1,4 +1,4 @@
-package bowling.application;
+package bowling.ui;
 
 import bowling.domain.state.Pin;
